refactor: split monolith modules, consolidate storage, remove dead code
Split annotation-popover.js (2256→684 lines) by extracting panel
builders, wirers, and shared controls into popover-panels.js.
Split floating-toolbar.js (1807→1375 lines) by extracting
documentation/guide templates into toolbar-docs.js.

Remove storage fallback writes from api-bridge.js — all mutations
now go through background.js respecting the storage lock.
Remove duplicate theme tokens from theme-manager.js (already in
styles.js :host rules). Remove dead code: unused functions, icons,
constants.

Server: remove dead applyAnnotationsUpdate(), extract
filterByUrlPattern() utility, add max watcher limit (100), remove
unused node-persist dep and execAsync import.
