+++ annotations-server/package.json
@@ -38,8 +38,7 @@
     "chalk": "^5.3.0",
     "commander": "^11.1.0",
     "cors": "^2.8.5",
-    "express": "^4.18.2",
-    "node-persist": "^3.1.3"
+    "express": "^4.18.2"
   },
   "repository": {
     "type": "git",
