+++ extension/content/modules/floating-toolbar.js
@@ -35,12 +35,6 @@ var VibeToolbar = (() => {
     collapse: '<svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><polyline points="9 18 15 12 9 6"/></svg>',
     // Vibe logo — actual icon (set dynamically in buildToolbar)
     logo: '',
-    eyeOff: '<svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M17.94 17.94A10.07 10.07 0 0 1 12 20c-7 0-11-8-11-8a18.45 18.45 0 0 1 5.06-5.94"/><path d="M9.9 4.24A9.12 9.12 0 0 1 12 4c7 0 11 8 11 8a18.5 18.5 0 0 1-2.16 3.19"/><path d="M1 1l22 22"/><path d="M14.12 14.12a3 3 0 1 1-4.24-4.24"/></svg>',
-    power: '<svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M18.36 6.64a9 9 0 1 1-12.73 0"/><line x1="12" y1="2" x2="12" y2="12"/></svg>',
-    // Theme icons
-    sun: '<svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2"><circle cx="12" cy="12" r="5"/><line x1="12" y1="1" x2="12" y2="3"/><line x1="12" y1="21" x2="12" y2="23"/><line x1="4.22" y1="4.22" x2="5.64" y2="5.64"/><line x1="18.36" y1="18.36" x2="19.78" y2="19.78"/><line x1="1" y1="12" x2="3" y2="12"/><line x1="21" y1="12" x2="23" y2="12"/><line x1="4.22" y1="19.78" x2="5.64" y2="18.36"/><line x1="18.36" y1="5.64" x2="19.78" y2="4.22"/></svg>',
-    moon: '<svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2"><path d="M21 12.79A9 9 0 1 1 11.21 3 7 7 0 0 0 21 12.79z"/></svg>',
-    system: '<svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2"><rect x="2" y="3" width="20" height="14" rx="2"/><line x1="8" y1="21" x2="16" y2="21"/><line x1="12" y1="17" x2="12" y2="21"/></svg>',
     // Links
     github: '<svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2"><path d="M9 19c-5 1.5-5-2.5-7-3m14 6v-3.87a3.37 3.37 0 0 0-.94-2.61c3.14-.35 6.44-1.54 6.44-7A5.44 5.44 0 0 0 20 4.77 5.07 5.07 0 0 0 19.91 1S18.73.65 16 2.48a13.38 13.38 0 0 0-7 0C6.27.65 5.09 1 5.09 1A5.07 5.07 0 0 0 5 4.77a5.44 5.44 0 0 0-1.5 3.78c0 5.42 3.3 6.61 6.44 7A3.37 3.37 0 0 0 9 18.13V22"/></svg>',
     server: '<svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><rect x="2" y="2" width="20" height="8" rx="2"/><rect x="2" y="14" width="20" height="8" rx="2"/><line x1="6" y1="6" x2="6.01" y2="6"/><line x1="6" y1="18" x2="6.01" y2="18"/></svg>',
@@ -62,16 +56,6 @@ var VibeToolbar = (() => {
     book: '<svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M4 19.5v-15A2.5 2.5 0 0 1 6.5 2H20v20H6.5a2.5 2.5 0 0 1 0-5H20"/></svg>',
     eye: '<svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M1 12s4-8 11-8 11 8 11 8-4 8-11 8-11-8-11-8z"/><circle cx="12" cy="12" r="3"/></svg>'
   };
-
-  const THEME_ICONS = { light: ICONS.sun, dark: ICONS.moon, system: ICONS.system };
-  const THEME_CYCLE = ['light', 'dark', 'system'];
-
-  function isLocalhost() {
-    const h = window.location.hostname;
-    return h === 'localhost' || h === '127.0.0.1' || h === '0.0.0.0'
-      || h.endsWith('.local') || h.endsWith('.test') || h.endsWith('.localhost')
-      || window.location.protocol === 'file:';
-  }
 
   async function init() {
     const root = VibeShadowHost.getRoot();
@@ -637,423 +621,13 @@ var VibeToolbar = (() => {
     }, 0);
   }
 
+  // Documentation/guide pages are in toolbar-docs.js (VibeToolbarDocs)
+  const _docsCallbacks = {
+    onBack: () => { closeSettings(); openSettings(); }
+  };
+
   function showDocumentation() {
-    if (!settingsDropdown) return;
-    const header = settingsDropdown.querySelector('.vibe-settings-header');
-    const body = settingsDropdown.querySelector('.vibe-settings-body');
-    if (!header || !body) return;
-
-    const version = chrome.runtime.getManifest().version;
-
-    // Replace header with back navigation
-    header.innerHTML = `
-      <button class="vibe-guide-back-btn" type="button" style="display:flex;align-items:center;gap:6px;background:none;border:none;cursor:pointer;color:var(--v-text-secondary);font-family:var(--v-font);font-size:13px;padding:0;">
-        ${ICONS.back}
-        <span style="color:var(--v-text-primary);font-weight:600;">Documentation</span>
-      </button>
-    `;
-
-    // Replace body with documentation links
-    body.innerHTML = `
-      <button class="vibe-settings-link vibe-get-started-guide-btn" type="button">
-        ${ICONS.rocket}
-        <span>Get started</span>
-        <span style="margin-left:auto;color:var(--v-text-secondary);">${ICONS.chevronRight}</span>
-      </button>
-      <button class="vibe-settings-link vibe-mcp-server-btn" type="button">
-        ${ICONS.server}
-        <span>MCP Server</span>
-        <span style="margin-left:auto;color:var(--v-text-secondary);">${ICONS.chevronRight}</span>
-      </button>
-      <div class="vibe-settings-separator"></div>
-      <button class="vibe-settings-link vibe-workflow-btn" data-workflow="single-page" type="button">
-        ${ICONS.webpage}
-        <span>Editing a single page</span>
-        <span style="margin-left:auto;color:var(--v-text-secondary);">${ICONS.chevronRight}</span>
-      </button>
-      <button class="vibe-settings-link vibe-workflow-btn" data-workflow="multi-page" type="button">
-        ${ICONS.globe}
-        <span>Editing multiple pages</span>
-        <span style="margin-left:auto;color:var(--v-text-secondary);">${ICONS.chevronRight}</span>
-      </button>
-      <button class="vibe-settings-link vibe-workflow-btn" data-workflow="collaborate" type="button">
-        ${ICONS.users}
-        <span>Collaborating</span>
-        <span style="margin-left:auto;color:var(--v-text-secondary);">${ICONS.chevronRight}</span>
-      </button>
-      <button class="vibe-settings-link vibe-workflow-btn" data-workflow="agents" type="button">
-        ${ICONS.robot}
-        <span>Annotating with agents</span>
-        <span style="margin-left:auto;color:var(--v-text-secondary);">${ICONS.chevronRight}</span>
-      </button>
-      <button class="vibe-settings-link vibe-workflow-btn" data-workflow="watch-mode" type="button">
-        ${ICONS.eye}
-        <span>Watch mode</span>
-        <span style="margin-left:auto;color:var(--v-text-secondary);">${ICONS.chevronRight}</span>
-      </button>
-      <div class="vibe-settings-separator"></div>
-      <a href="https://github.com/RaphaelRegnier/vibe-annotations" target="_blank" rel="noopener" class="vibe-settings-link">
-        ${ICONS.github}
-        <span>Contribute to Vibe Annotations</span>
-      </a>
-      <a href="https://github.com/RaphaelRegnier/vibe-annotations/releases/tag/v${escapeHTML(version)}" target="_blank" rel="noopener" class="vibe-settings-link">
-        ${ICONS.newspaper}
-        <span>Release notes</span>
-      </a>
-    `;
-
-    // Back button — restores full settings
-    header.querySelector('.vibe-guide-back-btn').addEventListener('click', () => {
-      closeSettings();
-      openSettings();
-    });
-
-    // Get started guide
-    body.querySelector('.vibe-get-started-guide-btn').addEventListener('click', () => showGetStartedGuide());
-
-    // MCP Server setup
-    body.querySelector('.vibe-mcp-server-btn').addEventListener('click', () => showWorkflow('mcp-setup'));
-
-    // Workflow navigation buttons
-    body.querySelectorAll('.vibe-workflow-btn').forEach(btn => {
-      btn.addEventListener('click', () => showWorkflow(btn.dataset.workflow));
-    });
-  }
-
-  function showGetStartedGuide() {
-    if (!settingsDropdown) return;
-    const header = settingsDropdown.querySelector('.vibe-settings-header');
-    const body = settingsDropdown.querySelector('.vibe-settings-body');
-    if (!header || !body) return;
-
-    header.innerHTML = `
-      <button class="vibe-guide-back-btn" type="button" style="display:flex;align-items:center;gap:6px;background:none;border:none;cursor:pointer;color:var(--v-text-secondary);font-family:var(--v-font);font-size:13px;padding:0;">
-        ${ICONS.back}
-        <span style="color:var(--v-text-primary);font-weight:600;">Get started</span>
-      </button>
-    `;
-
-    body.innerHTML = `
-      <div class="vibe-guide">
-        <div class="vibe-guide-section">
-          <div class="vibe-guide-label">1. Start annotating</div>
-          <p class="vibe-guide-text">Click the <strong>pencil button</strong> or your configured hotkey to enter inspection mode. Click any element to add a comment or modify its design.</p>
-        </div>
-
-        <div class="vibe-guide-section">
-          <div class="vibe-guide-label">2. Send to your agent</div>
-          <p class="vibe-guide-text">Hit <strong>Copy</strong> in the toolbar and paste into any AI chat, or <strong>Export</strong> to share a file. No server needed.</p>
-        </div>
-
-        <div class="vibe-guide-section">
-          <div class="vibe-guide-label">3. Install MCP server <span style="font-weight:400;color:var(--v-text-secondary);">(optional)</span></div>
-          <p class="vibe-guide-text">Let your coding agent fetch and resolve annotations automatically.</p>
-          <div class="vibe-guide-cmd" data-cmd="npm install -g vibe-annotations-server">
-            <code>npm install -g vibe-annotations-server</code>
-            <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-          </div>
-          <div class="vibe-guide-cmd" data-cmd="vibe-annotations-server start">
-            <code>vibe-annotations-server start</code>
-            <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-          </div>
-          <p class="vibe-guide-text" style="margin-top:8px;">Then connect your agent:</p>
-          <div class="vibe-guide-tabs">
-            <button class="vibe-guide-tab active" data-tab="claude">Claude Code</button>
-            <button class="vibe-guide-tab" data-tab="cursor">Cursor</button>
-            <button class="vibe-guide-tab" data-tab="windsurf">Windsurf</button>
-            <button class="vibe-guide-tab" data-tab="codex">Codex</button>
-            <button class="vibe-guide-tab" data-tab="openclaw">OpenClaw</button>
-          </div>
-          <div class="vibe-guide-panel active" data-panel="claude">
-            <div class="vibe-guide-cmd" data-cmd="claude mcp add --transport http vibe-annotations http://127.0.0.1:3846/mcp">
-              <code>claude mcp add --transport http vibe-annotations http://127.0.0.1:3846/mcp</code>
-              <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-            </div>
-          </div>
-          <div class="vibe-guide-panel" data-panel="cursor">
-            <p class="vibe-guide-text">Add to <strong>.cursor/mcp.json</strong>:</p>
-            <div class="vibe-guide-cmd" data-cmd='{"mcpServers":{"vibe-annotations":{"url":"http://127.0.0.1:3846/mcp"}}}'>
-              <code>{"mcpServers":{"vibe-annotations":{"url":"http://127.0.0.1:3846/mcp"}}}</code>
-              <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-            </div>
-          </div>
-          <div class="vibe-guide-panel" data-panel="windsurf">
-            <p class="vibe-guide-text">Add to Windsurf MCP settings:</p>
-            <div class="vibe-guide-cmd" data-cmd='{"mcpServers":{"vibe-annotations":{"serverUrl":"http://127.0.0.1:3846/mcp"}}}'>
-              <code>{"mcpServers":{"vibe-annotations":{"serverUrl":"http://127.0.0.1:3846/mcp"}}}</code>
-              <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-            </div>
-          </div>
-          <div class="vibe-guide-panel" data-panel="codex">
-            <p class="vibe-guide-text">Add to <strong>~/.codex/config.toml</strong>:</p>
-            <div class="vibe-guide-cmd" data-cmd="[mcp_servers.vibe-annotations]&#10;url = &quot;http://127.0.0.1:3846/mcp&quot;">
-              <code>[mcp_servers.vibe-annotations] url = "..."</code>
-              <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-            </div>
-          </div>
-          <div class="vibe-guide-panel" data-panel="openclaw">
-            <p class="vibe-guide-text">Add to <strong>~/.openclaw/openclaw.json</strong>:</p>
-            <div class="vibe-guide-cmd" data-cmd='{"mcpServers":{"vibe-annotations":{"url":"http://127.0.0.1:3846/mcp"}}}'>
-              <code>{"mcpServers":{"vibe-annotations":{"url":"http://127.0.0.1:3846/mcp"}}}</code>
-              <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-            </div>
-          </div>
-        </div>
-      </div>
-    `;
-
-    // Back → return to Documentation
-    header.querySelector('.vibe-guide-back-btn').addEventListener('click', () => showDocumentation());
-
-    // Tab switching
-    body.querySelectorAll('.vibe-guide-tab').forEach(tab => {
-      tab.addEventListener('click', () => {
-        body.querySelectorAll('.vibe-guide-tab').forEach(t => t.classList.remove('active'));
-        body.querySelectorAll('.vibe-guide-panel').forEach(p => p.classList.remove('active'));
-        tab.classList.add('active');
-        body.querySelector(`[data-panel="${tab.dataset.tab}"]`).classList.add('active');
-      });
-    });
-
-    // Copy buttons
-    body.querySelectorAll('.vibe-guide-copy').forEach(btn => {
-      btn.addEventListener('click', async () => {
-        const cmd = btn.closest('.vibe-guide-cmd').dataset.cmd;
-        await navigator.clipboard.writeText(cmd);
-        btn.innerHTML = ICONS.check;
-        setTimeout(() => { btn.innerHTML = ICONS.clipboard; }, 1500);
-      });
-    });
-  }
-
-  function showWorkflow(type) {
-    if (!settingsDropdown) return;
-    const header = settingsDropdown.querySelector('.vibe-settings-header');
-    const body = settingsDropdown.querySelector('.vibe-settings-body');
-    if (!header || !body) return;
-
-    const workflows = {
-      'mcp-setup': {
-        title: 'MCP Server',
-        content: `
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">1. Install and start the server</div>
-            <div class="vibe-guide-cmd" data-cmd="npm install -g vibe-annotations-server">
-              <code>npm install -g vibe-annotations-server</code>
-              <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-            </div>
-            <div class="vibe-guide-cmd" data-cmd="vibe-annotations-server start">
-              <code>vibe-annotations-server start</code>
-              <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-            </div>
-          </div>
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">2. Connect your agent</div>
-            <div class="vibe-guide-tabs">
-              <button class="vibe-guide-tab active" data-tab="claude">Claude Code</button>
-              <button class="vibe-guide-tab" data-tab="cursor">Cursor</button>
-              <button class="vibe-guide-tab" data-tab="windsurf">Windsurf</button>
-              <button class="vibe-guide-tab" data-tab="codex">Codex</button>
-              <button class="vibe-guide-tab" data-tab="openclaw">OpenClaw</button>
-            </div>
-            <div class="vibe-guide-panel active" data-panel="claude">
-              <div class="vibe-guide-cmd" data-cmd="claude mcp add --transport http vibe-annotations http://127.0.0.1:3846/mcp">
-                <code>claude mcp add --transport http vibe-annotations http://127.0.0.1:3846/mcp</code>
-                <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-              </div>
-            </div>
-            <div class="vibe-guide-panel" data-panel="cursor">
-              <p class="vibe-guide-text">Add to <strong>.cursor/mcp.json</strong>:</p>
-              <div class="vibe-guide-cmd" data-cmd='{"mcpServers":{"vibe-annotations":{"url":"http://127.0.0.1:3846/mcp"}}}'>
-                <code>{"mcpServers":{"vibe-annotations":{"url":"http://127.0.0.1:3846/mcp"}}}</code>
-                <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-              </div>
-            </div>
-            <div class="vibe-guide-panel" data-panel="windsurf">
-              <p class="vibe-guide-text">Add to Windsurf MCP settings:</p>
-              <div class="vibe-guide-cmd" data-cmd='{"mcpServers":{"vibe-annotations":{"serverUrl":"http://127.0.0.1:3846/mcp"}}}'>
-                <code>{"mcpServers":{"vibe-annotations":{"serverUrl":"http://127.0.0.1:3846/mcp"}}}</code>
-                <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-              </div>
-            </div>
-            <div class="vibe-guide-panel" data-panel="codex">
-              <p class="vibe-guide-text">Add to <strong>~/.codex/config.toml</strong>:</p>
-              <div class="vibe-guide-cmd" data-cmd="[mcp_servers.vibe-annotations]&#10;url = &quot;http://127.0.0.1:3846/mcp&quot;">
-                <code>[mcp_servers.vibe-annotations] url = "..."</code>
-                <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-              </div>
-            </div>
-            <div class="vibe-guide-panel" data-panel="openclaw">
-              <p class="vibe-guide-text">Add to <strong>~/.openclaw/openclaw.json</strong>:</p>
-              <div class="vibe-guide-cmd" data-cmd='{"mcpServers":{"vibe-annotations":{"url":"http://127.0.0.1:3846/mcp"}}}'>
-                <code>{"mcpServers":{"vibe-annotations":{"url":"http://127.0.0.1:3846/mcp"}}}</code>
-                <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-              </div>
-            </div>
-          </div>
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">3. Watch mode <span style="font-weight:400;color:var(--v-text-secondary);">(hands-free)</span></div>
-            <p class="vibe-guide-text">Your agent can automatically pick up annotations as you drop them. Just tell it:</p>
-            <div class="vibe-guide-cmd" data-cmd="Start watching Vibe Annotations">
-              <code>Start watching Vibe Annotations</code>
-              <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-            </div>
-            <p class="vibe-guide-text" style="margin-top:8px;">The agent calls <code>watch_annotations</code> in a loop, implements each change, and deletes the annotation when done. An eye icon appears in the toolbar and on badges while watching. Click the eye to stop.</p>
-          </div>
-        `
-      },
-      'single-page': {
-        title: 'Editing a single page',
-        content: `
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Best for quick edits</div>
-            <p class="vibe-guide-text">For a few annotations on one page, <strong>copy & paste</strong> is the fastest option. No server, no setup.</p>
-          </div>
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Workflow</div>
-            <p class="vibe-guide-text">1. Annotate elements on the page (comments, CSS tweaks, text changes)</p>
-            <p class="vibe-guide-text">2. Click <strong>Copy</strong> in the toolbar</p>
-            <p class="vibe-guide-text">3. Paste into any AI chat (Claude, ChatGPT, Cursor...) and ask the agent to implement the changes</p>
-          </div>
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Tips</div>
-            <p class="vibe-guide-text">Enable <strong>Clear on copy</strong> in settings to auto-delete annotations after copying. Keeps things clean between iterations.</p>
-            <p class="vibe-guide-text">Each annotation includes the selector, your comment, element context, and any pending changes. The agent gets everything it needs to locate and edit the right code.</p>
-          </div>
-        `
-      },
-      'multi-page': {
-        title: 'Editing multiple pages',
-        content: `
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Best for cross-page changes</div>
-            <p class="vibe-guide-text">When you're annotating across multiple routes, the <strong>MCP server</strong> is preferable. Your coding agent can read and resolve annotations from all pages at once, without manual copy-paste per route.</p>
-          </div>
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Setup</div>
-            <div class="vibe-guide-cmd" data-cmd="npm install -g vibe-annotations-server">
-              <code>npm install -g vibe-annotations-server</code>
-              <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-            </div>
-            <div class="vibe-guide-cmd" data-cmd="vibe-annotations-server start">
-              <code>vibe-annotations-server start</code>
-              <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-            </div>
-            <p class="vibe-guide-text" style="margin-top:8px;">Then connect your agent (e.g. Claude Code):</p>
-            <div class="vibe-guide-cmd" data-cmd="claude mcp add --transport http vibe-annotations http://127.0.0.1:3846/mcp">
-              <code>claude mcp add --transport http vibe-annotations http://127.0.0.1:3846/mcp</code>
-              <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-            </div>
-          </div>
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Workflow</div>
-            <p class="vibe-guide-text">1. Navigate your app and annotate elements across as many routes as needed</p>
-            <p class="vibe-guide-text">2. Tell your agent: <em>"read vibe annotations and implement the changes"</em></p>
-            <p class="vibe-guide-text">3. The agent pulls all pending annotations via MCP, edits your source files, and deletes each one when done</p>
-          </div>
-        `
-      },
-      collaborate: {
-        title: 'Collaborating with annotations',
-        content: `
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Annotations as a feedback tool</div>
-            <p class="vibe-guide-text">Anyone can annotate a website: add comments, tweak styles, edit text. Then <strong>export</strong> the annotations as a .json file and share it with a teammate.</p>
-          </div>
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Workflow</div>
-            <p class="vibe-guide-text">1. A reviewer annotates the live site (staging, production, or localhost)</p>
-            <p class="vibe-guide-text">2. They click <strong>Export</strong> and share the .json file (Slack, email, etc.)</p>
-            <p class="vibe-guide-text">3. A developer clicks <strong>Import</strong> on their localhost. Annotations, badges, and style previews appear instantly.</p>
-            <p class="vibe-guide-text">4. The developer copies or uses MCP to send the annotations to their coding agent</p>
-          </div>
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Cross-origin remap</div>
-            <p class="vibe-guide-text">Importing annotations from a public URL into localhost? The extension offers to <strong>remap URLs</strong> automatically so annotations anchor to your local dev server.</p>
-          </div>
-        `
-      },
-      'watch-mode': {
-        title: 'Watch mode',
-        content: `
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Hands-free annotation processing</div>
-            <p class="vibe-guide-text">Your coding agent automatically picks up and implements annotations as you drop them. No copy-paste, no manual triggering. Requires the MCP server.</p>
-          </div>
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Start watching</div>
-            <p class="vibe-guide-text">Tell your agent:</p>
-            <div class="vibe-guide-cmd" data-cmd="Start watching Vibe Annotations">
-              <code>Start watching Vibe Annotations</code>
-              <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-            </div>
-            <p class="vibe-guide-text" style="margin-top:8px;">The agent calls <code>watch_annotations</code> in a loop, implements each change, and deletes the annotation when done.</p>
-          </div>
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Visual feedback</div>
-            <p class="vibe-guide-text">While watching, an eye icon replaces the copy button in the toolbar, and badges show eyes instead of numbers. Click the eye to stop watching.</p>
-            <p class="vibe-guide-text">Auto-stops after 5 minutes of inactivity.</p>
-          </div>
-        `
-      },
-      agents: {
-        title: 'Annotating with agents',
-        content: `
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Let agents annotate for you</div>
-            <p class="vibe-guide-text">Agents can help you annotate collaboratively, or work fully autonomously to review any site.</p>
-          </div>
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Compatible agents</div>
-            <p class="vibe-guide-text"><strong>Claude Chrome extension</strong> has direct page access and can call the API from its javascript tool.</p>
-            <p class="vibe-guide-text"><strong>OpenClaw</strong> uses CDP evaluate to run JS on the page.</p>
-            <p class="vibe-guide-text"><strong>Claude Code, Cursor, Windsurf</strong> can access the page via a DevTools MCP server or Playwright.</p>
-          </div>
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Prompt to get started</div>
-            <p class="vibe-guide-text">Copy this and paste it into your agent's chat to orient it towards the bridge API:</p>
-            <div class="vibe-guide-cmd" data-cmd="Read window.__vibeAnnotations.help() and use this extension for my comments on this project.">
-              <code>Read window.__vibeAnnotations.help() and use this extension for my comments on this project.</code>
-              <button class="vibe-guide-copy" type="button">${ICONS.clipboard}</button>
-            </div>
-          </div>
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">Requirement</div>
-            <p class="vibe-guide-text">The extension must be active on the page for the bridge API to be available. This works best when the agent uses <strong>your browser</strong> (Claude Chrome, DevTools MCP), since the extension is already installed.</p>
-            <p class="vibe-guide-text">Agents that launch their own browser (Playwright, Puppeteer) won't have the extension loaded by default. This can be configured by passing the extension path at launch, but requires some local setup.</p>
-          </div>
-          <div class="vibe-guide-section">
-            <div class="vibe-guide-label">How it works</div>
-            <p class="vibe-guide-text">The agent calls <code>__vibeAnnotations.help()</code> to discover the API, then uses <strong>createStyleAnnotation</strong> for broad CSS changes and <strong>createAnnotation</strong> for single-element edits. Changes preview live in the browser and get recorded as annotations for a coding agent to implement in source.</p>
-          </div>
-        `
-      }
-    };
-
-    const wf = workflows[type];
-    if (!wf) return;
-
-    header.innerHTML = `
-      <button class="vibe-guide-back-btn" type="button" style="display:flex;align-items:center;gap:6px;background:none;border:none;cursor:pointer;color:var(--v-text-secondary);font-family:var(--v-font);font-size:13px;padding:0;">
-        ${ICONS.back}
-        <span style="color:var(--v-text-primary);font-weight:600;">${wf.title}</span>
-      </button>
-    `;
-
-    body.innerHTML = `<div class="vibe-guide">${wf.content}</div>`;
-
-    // Back → return to Documentation
-    header.querySelector('.vibe-guide-back-btn').addEventListener('click', () => showDocumentation());
-
-    // Copy buttons (for MCP workflow)
-    body.querySelectorAll('.vibe-guide-copy').forEach(btn => {
-      btn.addEventListener('click', async () => {
-        const cmd = btn.closest('.vibe-guide-cmd').dataset.cmd;
-        await navigator.clipboard.writeText(cmd);
-        btn.innerHTML = ICONS.check;
-        setTimeout(() => { btn.innerHTML = ICONS.clipboard; }, 1500);
-      });
-    });
+    VibeToolbarDocs.showDocumentation(settingsDropdown, ICONS, _docsCallbacks);
   }
 
   function closeSettings() {
@@ -1217,10 +791,6 @@ var VibeToolbar = (() => {
       updateUI();
       VibeEvents.emit('watch:changed', { active: watcherActive });
     }
-  }
-
-  function showCopyFeedback() {
-    // Will be used by View All panel copy button
   }
 
   function animateToolbarOut() {
@@ -1617,8 +1187,6 @@ var VibeToolbar = (() => {
     div.textContent = str;
     return div.innerHTML;
   }
-
-  function capitalize(s) { return s.charAt(0).toUpperCase() + s.slice(1); }
 
   function formatShortcut(sc) {
     const parts = [];
