+++ extension/manifest.json
@@ -65,7 +65,9 @@
         "content/modules/element-context.js",
         "content/modules/badge-manager.js",
         "content/modules/inspection-mode.js",
+        "content/modules/popover-panels.js",
         "content/modules/annotation-popover.js",
+        "content/modules/toolbar-docs.js",
         "content/modules/floating-toolbar.js",
         "content/modules/bridge-handler.js",
         "content/content.js"
